% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollarScraper.R
\name{CollarScraper-package}
\alias{CollarScraper-package}
\alias{CollarScraper}
\title{An R scraper for downloading GPS collar data.}
\description{
These are scrapers for downlading GPS collar
data from various vendor websites.
They are based on logging into and parsing their web-based
user interfaces.
}
\references{
http://github.com/rgzn/CollarScraper
}
\author{
Jon Weissman
}
